// src/codepage/tvm-spec.d.ts
export type InstructionName = string;
/**
 * Global version (ConfigParam 8) which enables this instruction. Version 9999 means that instruction has no global version and currently unavailable in mainnet.
 */
export type SinceGlobalVersion = number;
/**
 * Free-form bytecode format description.
 */
export type OpcodeFormatDocumentation = string;
/**
 * Free-form description of stack inputs and outputs. Usually the form is `[inputs] - [outputs]` where `[inputs]` are consumed stack values and `outputs` are produced stack values (top of stack is the last value).
 */
export type StackUsageDescription = string;
export type CategoryOfInstruction = string;
/**
 * Free-form markdown description of instruction.
 */
export type InstructionDescription = string;
/**
 * Free-form description of gas amount used by instruction.
 */
export type GasUsageInfo = string;
/**
 * Free-form fift usage description.
 */
export type FiftUsageDoc = string;
export type FiftSnippet = string;
export type ExampleDescription = string;
/**
 * TL-b bytecode format description.
 */
export type TLBSchema = string;
/**
 * Prefix to determine next instruction to parse. It is a hex bitstring as in TL-b (suffixed with `_` if bit length is not divisible by 4, trailing `'1' + '0' * x` must be removed).
 */
export type InstructionPrefix = string;
// ...
export type InstructionOperands = Operand[];
// ...
export type PossibleBranchesOfAnInstruction = Continuation[];
/**
 * Can this instruction not perform any of specified branches in certain cases (do not modify cc)?
 */
export type NoBranchPossibility = boolean;
// ...
export interface Instruction {
  mnemonic: InstructionName;
  since_version: SinceGlobalVersion;
  doc: Documentation;
  bytecode: BytecodeFormat;
  value_flow: ValueFlowOfInstruction;
  control_flow: ControlFlowOfInstruction;
}
/**
 * Free-form human-friendly information which should be used for documentation purposes only.
 */
export interface Documentation {
  opcode?: OpcodeFormatDocumentation;
  stack?: StackUsageDescription;
  category: CategoryOfInstruction;
  description: InstructionDescription;
  gas: GasUsageInfo;
  fift: FiftUsageDoc;
  fift_examples: {
    fift: FiftSnippet;
    description: ExampleDescription;
  }[];
}
/**
 * Information related to bytecode format of an instruction. Assuming that each instruction has format `prefix || operand_1 || operand_2 || ...` (also some operands may be refs, not bitstring part).
 */
export interface BytecodeFormat {
  tlb: TLBSchema;
  prefix: InstructionPrefix;
  operands_range_check?: OperandsRangeCheck;
  operands: InstructionOperands;
}
/**
 * In TVM, it is possible for instructions to have overlapping prefixes, so to determine actual instruction it is required to read next `length` bits after prefix as uint `i` and check `from <= i <= to`. Optional, there is no operands check in case of absence.
 */
export interface OperandsRangeCheck {
  length: number;
  from: number;
  to: number;
}
/**
 * Information related to usage of stack and registers by instruction.
 */
export interface ValueFlowOfInstruction {
  inputs: InstructionInputs;
  outputs: InstructionOutputs;
}
/**
 * Incoming values constraints.
 */
export interface InstructionInputs {
  stack?: StackValues;
  registers: RegisterValues;
}
export interface MatchArm {
  value: ArmValue;
  stack: StackValues;
}
/**
 * Outgoing values constraints.
 */
export interface InstructionOutputs {
  stack?: StackValues;
  registers: RegisterValues;
}
/**
 * Information related to current cc modification by instruction
 */
export interface ControlFlowOfInstruction {
  branches: PossibleBranchesOfAnInstruction;
  nobranch: NoBranchPossibility;
}